Move components into tutorial directories
