--- website/src/theme/ReactLiveScope/queries.js
@@ -1,21 +1,0 @@
-import { gql } from 'graphql-request';
-
-export async function publish(client, entry, operation) {
-  const query = gql`
-    mutation Publish($entry: String!, $operation: String!) {
-      publish(entry: $entry, operation: $operation) {
-        logId
-        seqNum
-        backlink
-        skiplink
-      }
-    }
-  `;
-
-  const result = await client.request(query, {
-    entry,
-    operation,
-  });
-
-  return result.publish;
-}
